update persistent text field
