+++ lib/basf_flutter_components.dart
@@ -15,6 +15,8 @@ export 'src/widgets/buttons/outlined_button.dart';
 export 'src/widgets/buttons/slider_button.dart';
 export 'src/widgets/buttons/text_button.dart';
 export 'src/widgets/dialogs.dart';
+export 'src/widgets/inputs/logic/persisted_data.dart';
+export 'src/widgets/inputs/logic/persisted_input_cubit.dart';
 export 'src/widgets/inputs/persisted_text_field.dart';
 export 'src/widgets/inputs/text_field.dart';
 export 'src/widgets/options/checkboxes.dart';
